Create SortIcon.tsx
